requirements.txt
fastapi
uvicorn[standard]
Jinja2
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy
pydantic
python-multipart
aiofiles